MarketData: restructure data retrieval
