+++ rubi/rubi/rubicon_types/__init__.py
@@ -1,3 +1,5 @@
 from .order import *
 from .orderbook import *
 from .pair import *
+
+from .order_query import *
